Move NSProxy to `objc2::runtime` internally
So that we need to rely on fewer macro internals in `icrate`
