+++ crates/icrate/tests/proxy.rs
@@ -1,0 +1,8 @@
+#![cfg(feature = "Foundation_NSProxy")]
+use icrate::Foundation::NSProxy;
+use objc2::ClassType;
+
+#[test]
+fn dummy() {
+    let _cls = NSProxy::class();
+}
