+++ crates/objc2/src/runtime/nsobject.rs
@@ -59,10 +59,12 @@ unsafe impl ClassType for NSObject {
         }
     }
 
+    #[inline]
     fn as_super(&self) -> &Self::Super {
         &self.__inner
     }
 
+    #[inline]
     fn as_super_mut(&mut self) -> &mut Self::Super {
         &mut self.__inner
     }
@@ -208,6 +210,7 @@ impl hash::Hash for NSObject {
 }
 
 impl fmt::Debug for NSObject {
+    #[inline]
     #[doc(alias = "description")]
     #[doc(alias = "debugDescription")]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
