+++ crates/test-ui/ui/msg_send_id_invalid_return.stderr
@@ -29,6 +29,7 @@ error[E0277]: the trait bound `objc2::runtime::Class: Message` is not satisfied
             Exception
             NSObject
             ProtocolObject<P>
+            __NSProxy
             __RcTestObject
             objc2::runtime::Object
   = note: required for `RetainSemantics<1>` to implement `MsgSendId<&objc2::runtime::Class, Id<objc2::runtime::Class>>`
@@ -46,6 +47,7 @@ error[E0277]: the trait bound `objc2::runtime::Class: Message` is not satisfied
             Exception
             NSObject
             ProtocolObject<P>
+            __NSProxy
             __RcTestObject
             objc2::runtime::Object
   = note: required for `RetainSemantics<1>` to implement `MsgSendId<&objc2::runtime::Class, Id<objc2::runtime::Class>>`
@@ -81,6 +83,7 @@ error[E0277]: the trait bound `objc2::runtime::Class: Message` is not satisfied
             Exception
             NSObject
             ProtocolObject<P>
+            __NSProxy
             __RcTestObject
             objc2::runtime::Object
   = note: required for `RetainSemantics<2>` to implement `MsgSendId<&objc2::runtime::Class, Allocated<objc2::runtime::Class>>`
