+++ crates/test-ui/ui/protocol.stderr
@@ -8,6 +8,7 @@ error[E0277]: the trait bound `dyn Foo: NSObjectProtocol` is not satisfied
             DummyClass
             NSObject
             ProtocolObject<T>
+            __NSProxy
   = note: required for `ProtocolObject<dyn Foo>` to implement `NSObjectProtocol`
 note: required by a bound in `impl_nsobject`
  --> ui/protocol.rs
